Updated InformationPostingViewController.
- Added initial layout references: labels and textfield.
- The InformationPostingViewController now includes a text field
where the location information is supposed to be entered.

- Added CommonTextFieldDelegate class.
This class is supposed to handle all text field inputs in a
centralised manner. This includes for instance ending the
editing of text when the enter button is pressed on the iOS
keyboard.

- When the button at the bottom is pressed (to find the location
on the map), the labels are automatically hidden.