Added student request parameters.
When requesting the student location data from the ParseAPI servers,
the request URL now uses parameters which limit the amount of locations
to 100 and requests that the response order is sorted by update time of
the student locations. In other words, the last 100 posted locations
are received in order of which they have been posted to the ParseAPI
servers.